feat: adds Digits constraint

+++ src/main/java/br/com/fiap/autoatendimento/dataprovider/api/dto/request/DadosPedidoPagamentoReqDto.java
@@ -11,6 +11,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -27,7 +29,7 @@ public class DadosPedidoPagamentoReqDto {
     private String notificationUrl;
 
     @JsonProperty("total_amount")
-    private Double totalAmount;
+    private BigDecimal totalAmount;
 
     @JsonProperty("description")
     private String description;
